[ADD] runbot_merge: fwnice mode
Opt-in to allow creating forward ports in `nice` priority. Allows
prioritising (heh) actual PRs in slots-constrained times e.g. during
freezes and the like.

Both PRs and forwardports can still be reprioritised normally, and
this downgrading only applies to forward ports of PRs in default
mode. Note that if a forward-port is upgraded to `default`, this
upgrade will stick to *its* forward-ports.

Fixes #1211