[IMP] runbot_merge: staging consistency from a split
Stagings have a fairly consistent ordering. When looking at splits
it's convenient that their batches follow the ordering of the source
staging, that makes it easier to know what PR went where and
importantly which PR disappeared if any.

However as it turns out the second part ever happening was sheer luck,
as the split just has an o2m to the batches, so the order in which
batches are moved from a split to a staging is arbitrary and down to
the order of iteration of the values in the `split_id` index (since we
have a fixed `split_id`).

It should be possible to mostly mitigate that issue by sorting the
batches with the same key we're using to look them up for
staging. This may not be 100% as it's technically possible for the key
to be updated between the initial staging and the split's
staging (probably mostly by setting the priority) but it should work
in 99% of cases.