[IMP] runbot_merge: surface batches removed from split
Before this change, batches removed from a split between its creation
and its re-staging (whether because of r-, being closed, because of an
optional status failure, ...) pretty much just disappear which can
make some staging failure forensics more difficult than necessary to
track (e.g. did the splits succeed because it was a nondet error or
because the offending PR was removed?).

Add tracking of "original batches" through splits and re-staging. Note
that this is performed via unstructured json fields, and currently
does not track down to the level of PR either so if a batch has
multiple PRs, gets staged, fails, and one of the PR is closed, that
will not be surfaced (for now?).

Fixes #1216