[IMP] runbot_merge: log info when skipping a PR
Might even be worth posting a message. It looks like `mergeable` is a
lot less stable than assumed especially as of late, as on this
2025-10-01 we have had two different splits where a bunch of PRs got
skipped due to this flag... or so I assume: the logs say

> Staging pr <pr> for target <branch>; method=<method>

then nothing, whereas normally this should be followed by

> rebasing <pr> on <base>
> copied <commit> to <commit> (parent: <commit>)
> Staged PR <pr> to <branch> by <method>: <head> -> <staged_head>

or some sort of error. This strongly hints that the staging was
skip-ed, which currently only occurs for mergeability reasons.