[IMP] runbot_merge: sort read PRs in priority order
Makes it easier to see what order they'll be staged in.

Also update the batch key to properly handle sorting batches which
are *not* unblocked, just in case (actually mostly because I initially
wanted to sort the entire set of ready PRs, but then I realised it was
not super useful to order PRs which are in error or blocked).