[FIX] runbot: avoid being stuck when dropping local db
In some rare case the dropdb can be stuck. Even if incomplete the latest
investigation revealed that it can happen in rare case when a listen bus
loop is stuck by the GIL in one build.

Adding a timeout on the dropdb should mitigate this issue by skipping
the dropdb and hoppefuly killing the stuck build at some point.