[FIX] runbot: sum stats
Currently, we have different stat regexes for post-install, at install,
etc., to avoid stat category name clashes. When some category stats have
the same name and are present in the same build, only one of the values
will be taken, hiding the other one.

The recent work to simplify stat collection caused an issue here by
merging at install and post-install test stats.

This is also an issue with the web module, which is now run in all
subbuilds.

This new version proposes to sum the stats instead.

This can cause some problems too, since the install times are now doubled.
This problem is solved by allowing filtering of the stats with the config
step name.

While at it, the trigger selection logic is moved to the stat page
instead of a dropdown, and the layout is improved.

Performance is also improved by searching the batch and not the
build. This could result in fewer stats if some builds are included
in multiple batches, but the performance gain is worth it.