[REF] runbot: libraries should live in static/lib
Libraries used on the client side are usually stored in
`<mymodule>/static/lib`, leaving `<mymodule>/static/src` for the actual
codebase. This is expected by the `AssetsBundle` and `JavascriptAsset`.

The difference made between the two is that JavaScript files are
transpiled by default to Odoo modules when put in `static/src`.

This commit moves all the libraries to `static/lib` instead and reverts
commit 33bef243d20c8dc1213e086e655fbbb540599a28 because isn't needed
anymore.