[IMP] runbot: allow to send status in a service
Currently the leader and status service are synchronous, meaning that if
either fetching a repo or sending a status is slow, the other is
impacted.

In some cases, a fetch may take a verry long time pausing the status
updates.
In other cases, github is slow to answer making the satus updates slow
taking some times multiple minutes to be sent.
Moreover, during the send of a status batch, new status could already
exists making the sent status outdated.

To solve these issues, we introduce a status service that will be
responsible for sending the status updates. This is optional and by
default the leader will still send the status updates.

Additionnally, we only send status on build older than 2 minutes to
avoid sending status that could be quickly outdated (style, cla,
instantly failling builds).