[IMP] runbot: ignore autoretry tb
When parsing the logs to find traceback, runbot will catch errors that
should be hidden by autoretry

2025-12-03 01:22:00,501 26 _WARNING 94760443-saas-18-2-all odoo.tests.common: Error during browser shutdown
Traceback (most recent call last):
  File "/data/build/odoo/odoo/tests/common.py", line 1511, in _websocket_request
    return f.result(timeout=timeout)
           ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/concurrent/futures/_base.py", line 458, in result
    raise TimeoutError()
TimeoutError

The Traceback (most recent call last): will be parsed and make the build
fail.

This pr will detect if the previous line was a "autoretry" line by
checking if it contains a log level starting with _ and ignore traceback
in this case.

This was paritially solved in odoo by altering the log to alter the
traceback prefix but this does not cover all cases, and makes odoo code
more complex than needed.